Added "Run custom command" callback.
Added a callback to the check box for running a custom command at start.
The callback disables custom command entry if you're not running a custom
command.

Also, I moved all of the callbacks up to the top, so that we don't have to
worry about their order, and they are easier to find.