Mark lots of functions static.
This marks most (if not all) functions that can be static as such. This also
applies to global variables. This has the effect of making it less confusing
about what functions can be used where.

Also, I removed the goto_tab_* functions and replaced them with the function
goto_tab_generic(). Of course, I had to make wrapper functions to make the
callbacks work as expected, but at least all of the code is in one place now,
instead of being in 10.

Removed the unnecessary function clean_up_no_main(), since no-one was using
it.