Add config file versioning / updating system.
This adds a version string to the config file which tells you which version of
tilda created this config file.

Having this information, it is very easy to add custom things to update an
older configuration to the new standard. There is a template for how to add an
updater in the try_to_update_config_file() function.