Fix handling of the icon installation directory.
This (attempts) to follow the GNU standards for Autoconf,
which allow things like:
./configure --prefix=/usr
make
make install prefix=/opt

Which installs the package in a different prefix than was defined at configure
time. Kinda stupid if you ask me...

Note that tilda doesn't even try and detect the path to the icon. It uses the
prefix that is given at configure time only.