Switch the configuration wizard over to using glade
This implements the config wizard using glade instead of having it built by
hand. This makes it much easier to change the look and placement in the
dialog without changing the code.

NOTE: there are still a few bugs to work out. These are marked with FIXME
in the wizard.c code.

NOTE: GNU Gettext support is still missing.

Signed-off-by: Ira W. Snyder <devel@irasnyder.com>