Make WWW links open a web browser when clicked
This adds a patch (very slightly changed) from the Debian repository that
makes WWW links open in a web browser when you Ctrl+Left Click on them.
Currently, it spawns /usr/bin/x-www-browser. This will need to be made
configurable before release.

Thanks to catoblepa on IRC for sending me the patch!

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>