Fix Use Image for Background check box
The Use Image for Background check box was not always causing there not to
be an image displayed even when it was not checked. The "selection-changed"
property of the image selector was getting hit upon creating the
preferences window. I'm not sure why. This makes sure only to run the
handler if we need to.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>