Add update code for the key grabber
When we changed to tomboy's keygrabber, we had to change the format of the
keybinding. Even though we have a great new wizard to set the keybinding,
we still need to update for old users.

This adds a lex file that will do the conversion (it makes a really easy
parser). It also hooks it into the configuration upgrading system.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>