Major audit of tilda
This is the results of a major audit of the tilda code. I went through and
reorganized the codebase to make things work in a more sane order. This
involved creating _init() and _free() functions for the two major data
structures: tilda_window and tilda_term.

I also moved all functions that should have been local to the proper source
file and made them static. This mostly involved callbacks, which were moved
to the file where they were attached.

I renamed functions to have the same convention. This mostly applies to
callbacks, which now have _cb() on the end. There is likely more cleanup
yet to be done in this area.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>