Better Animation Support
The old way that we were doing animation was by forcing the main tilda
window to resize immediately. This worked decently, but setting the window
into this mode now causes crashes.

This method forces gtk to run the main loop during the middle of a
callback. It seems to have the good effect that the VteTerminal also
resizes nicely during the animation.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>