Fix close of wizard without pressing "Close"
The wizard window would do very bad things because the wizard_closed()
function was not being called when the X in the upper right was pressed,
instead of pressing the "Close" button.

This adds the necessary callback to the window's delete_event signal to fix
the problem. Now the wizard_closed() function is always called when the
wizard is closed.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>