[BUGFIX] The next_tab() and prev_tab() shortcuts make terminal scroll
This is based on the solution to the ^T bug which was recently fixed. Before
this fix, if you opened 2 or more tabs, filled them with stuff (for example,
dmesg) then used the next_tab() and prev_tab() keybindings, the terminal
would scroll upwards (or downwards). This happened because the keystrokes
were not stopped by the callback, but continued to propagate.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>