Update Documentation + Change Version to 0.9.6pre
This fixes some misspellings in the documentation, updates the README to
have the correct keybindings for Previous and Next tab, and also removes
some extraneous blanks at the end of lines in the ChangeLog.

I added my future plans for the next major Tilda release. The direction I'd
like to take the project in. I could really use some help with it. :) I'm
definitely no DBus master.

I also changed the version number to 0.9.6pre, since this is a pre-release,
after all.

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>