Remove use of deprecated GtkItemFactory in tilda_term
This removes the use of deprecated GTK+ widget GtkItemFactory, which was
used to create the popup menu upon right-click. The preferred method is to
use GtkUIManager.

While there is a little more code here, it does have the side effect of
removing a couple of translatable strings that we don't need, because
the strings associated with the stock items were exactly the same. :)

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>