Another try at thwarting focus-stealing prevention
It seems that the last change to try to address the focus-stealing prevention
issues in Metacity weren't quite enough. There were still some issues with
Tilda not getting the focus when it deserved it. This is another relatively
similar approach to the same problem. It still updates _NET_WM_USER_TIME, but
does it as early as possible in the pull() method.