[Key Grabber] Fix key grabber for metacity-2.22
With the release of GNOME 2.22, and therefore metacity-2.22, Tilda is
broken. It never gets the focus anymore.

This adds a huge amount of code to make metacity work again. It pretends
to be a pager, so that it gives us maximum permission to focus windows.

I left the original _NET_WM_USER_TIME hack in place, just in case, even
though the new code makes metacity work without it.

Tested in:
Metacity-2.20.2
Metacity-2.22.0
KWin-3.5.9
Fluxbox-1.0.0
Openbox-3.4.7_pre3

Signed-off-by: Ira W. Snyder <tilda@irasnyder.com>