Added '--enable-lto' option.
While nothing in tilda is in any way performance critical, link
time optimization may produce smaller binaries (in my case, the
savings are a bit more than 4K for either -O2 or -Os) and comes
basically for free (tested with gcc 4.8.2).

This option is not enabled by default, since it needs testing
and older compilers (GCC < 4.6) don't support link time
optimizations out of the box.

In the future, one may consider enabling the option by default
if a recent compiler and toolchain is found.