Don't set tilda active when sub-windows are open.
The wizard uses "disable_auto_hide" to say that its window is open.  If the
window is set active on mouseover when the settings window is open then the
settings window will be underneath the terminal and not be usable at all.

This reverts "1bd3bbffdac68cc48028c0eef6cbd1a034e0ece0"