wizard: migrate keybinding code to tilda-keybinding.h
This removes the keybinding code from the wizard and uses the functions
from tilda-keybinding.h

As a result we will now have a scrollable list of keybindings.
This replaces the old individual buttons for each keybinding.

This code was originally authored by x2b and refactored
by Alexander Maznev and Sebastian Geiger.

Fixes #163
Fixes #234

Signed-off-by: pik <alexander.maznev@gmail.com>
Signed-off-by: Sebastian Geiger <sbastig@gmx.net>