Fix copy/paste bug when scrollbar is on the left
This patch fixes a bug that caused the copy and paste
hotkeys to become broken, when the scrollbar was made visible
on the left. With the scrollbar on the right or disabled this
bug was not present.

The reason was that we looked up the vte_terminal widget
based on its location in its parent container and assumed the terminal
widget was always the first child of its parent. If the scrollbar is
on the left then this assumption was not true.

The solution is to lookup the VteTerminal widget from list of
tilda terminals, which keep a pointer to the VteTerminal widget.