Wizard: Avoid error when the wizard is opened multiple times
Before this patch we would log an error if the wizard was aready opened and the user tried
to open it again from the context menu. This patch fixes this issue. We now keep a pointer
to the wizard dialog in the tilda window structure and check for it when the wizard is opened.
If a wizard window already exists, then we just present it, otherwise we preceed as normally.

This patch moves the assignment to the ugly global tw pointer, which reminds me once again,
that this should be fixed (by registering the tw pointer as user data to all the callbacks).
But this patch is not the right scope for that.