Fix for segfault on opening wizard
When the wizard is opened via Right click > Preference, it causes a segfault due
to `wizard()` function receives a window object (`tw`) with `tw->wizard_window`
having 0xFFFFFFFF in it. Obviously this is not a valid reference to an existing
wizard window (there were none), but this value causes the code to assume there
is already a wizard window and trying to show it.

Fixed it by simply checking if the passed in `tw->wizard_window` is 0xFFFFFFFF
and setting it to NULL.