Search: Load search widgets from GtkBuilder
The search bar is now implemented in GtkBuilder and can be designed
and viewed with glade.

A few memory issues have been fixed.

Search direction and checkbox options are no longer dependend on the English
strings so they can be localized.

Added a tilda_search structure that holds the widgets from GtkBuilder.

The tilda_window struct now holds a reference to GtkBuilder that could
also be used by the Wizard in the future.

Showing and hiding the label to indicate that the search reached the bottom
of the file still needs some logic to toggle its visibility.