tilda_window.c: Keep tw->terms in sync with tab order
Without this patch functions which rely on the fact that the order of terminals in the terms structure and the order of tabs in the GtkNotebook
is the same did not function correctly if tabs were reordered.

This patch synchronizes the order of the tabs in the GtkNotebook and the order
in the tw->terms structure.

Fixes #173