libvte: Implement optional support for libvte-2.91
The vte version is now printed in the summary information
of the configure script and can be turned on or for with
`--enable-vte-2.91`.

This commit also adds a function to handle deprecated
values in libconfuse and remove them from the config
file.

Since libvte-2.91 deprecated and removed some options,
we are hidding those options in the wizard now.

Fixes #94
Fixes #138

Signed-off-by: Sebastian Geiger <sbastig@gmx.net>