Show main window before adding the first terminal tab
After commit 47c679f scollbar on the first terminal window/tab was showing even
when the scollbars are disabled in the settings.

Appears showing all widgets (for main window) override the hidden status of
scrollbar in the tab that was initializes earlier.

Moving part of the code that show the main window before adding the tab seem to
fix the issue.