Window: Simple keybinding for cycling monitor
Notes / loose ends:

* I went from an original approach of setting the relative x and y coordinates
to (0, 0) after cycling to the next monitor, to instead preserving the relative
x and y against the newly-selected display. In both cases not messing with the
size. This might not be the *best* approach -- some other possibilities I can
think of would be (1) a "proportional transposition", or (2) just setting the
position and/or size to arbitrary, consistent values (like filling the entire
workspace on the new monitor). I spend almost all my time in fullscreen mode,
however, so I am neither affected by nor the best judge of this use case.

* No effort is made to update the wizard UI if we cycle monitors while the
wizard is open. (It looks like we aren't supposed to be able to trigger hotkeys
while the wizard is open, but I was able to get it to work by changing focus.)
This decicison is based on the precedent set by other hotkeys such as "enable
fullscreen".

* What would an integration with the wip/monitornames branch look like? Or,
that branch, done right and completely, probably supersedes us.