Window: Logic for position disagreement
Sometimes -- usually if a monitor is hotplugged or the multiple-monitor
geometry is changed -- the prevailing (x_pos, y_pos) in the absolute
coordinate space do not agree with show_on_monitor_number. We now add some
additional logic to trust show_on_monitor_number above all; if the newly-
calculated absolute (x, y) that we get from trying to retain relative
position on our new monitor places the window outside the workarea of the
destination monitor, we drop the hammer and set x and/or y to 0 as needed.

This means, for example, if you have a Dell monitor on your left (as monitor
0) and an LG monitor on your right (as monitor 1), and have tilda on monitor
0 (the Dell), but then change your virtual monitor configuration such that the
LG is to the left of the Dell, tilda will go from being on monitor 0 to monitor
1, but without updating its configuration values. And your absolute coordinates
might be messed up, too. With this patch, next time you cycle the monitors,
you'll land tilda on the Dell (where it already is!), but after that, things
will operate normally.