tilda-keybinding: new keybinding logic
This patch introduces a new pair of files for the keybinding logic. Its
will populate the GtkTreeView in the wizard dialog and handles
the updates of the keybindings by showing a dialog that grabs
the new keybinding for a particular hotkey.

This code was originally authored by x2b and refactored
by Alexander Maznev and Sebastian Geiger.

Fixes #163
Fixes #234

Signed-off-by: pik <alexander.maznev@gmail.com>
Signed-off-by: Sebastian Geiger <sbastig@gmx.net>