wizard: add new TildaWizard structure
This is new structure is intended to keep variables that are only relevant for the
wizard. Currently we are keeping many varibles in the tilda_window structure,
which creates an unnecessary tight coupling with the tilda_window.

With this chance in place we will also be able to get rid of the global
GtkBuilder variable named 'xml'.