[search] migrate to PCRE2 regex handling
This patch replaces the deprecated VTE search functions
based on GRegex with the newer VteRegex-based functions
that use PCRE2.

This patch breaks on Ubuntu 18.04 and 18.10 which have
the required VTE version above 0.46. The reason is that
Ubuntu has patched PCRE2 out of VTE.

The following links provide additional information on
this matter:

* https://github.com/lanoxx/tilda/pull/399
* https://github.com/gnunn1/tilix/issues/916
* https://bugs.launchpad.net/ubuntu/+source/gnome-terminal/+bug/1666264