configsys: deprecate absolute size values
This deprecates the absolute sizes and replaces them with relative
sizes in percent in the tilda config.

The wizard UI is adjusted and will now show two fraction
digits for the percentage values.

Migration code is added to automatically migrate
the absolute size values to the new percentage values
using the current monitor workarea size.