configsys: set default palette index to 1
The default configuration values for the initial palette are identical to the
'tango' color palette, but we used to initialize the default 'palette_scheme' value to 0
which would then show 'Custom' rather than the tango color scheme in the wizard.

The correct value should be 1 so that the user sees 'Tango' as color scheme.