replace deprecated functions for monitor handling
This removes the deprecated GdkScreen functions for monitor
handling and replaces then with the respective GdkMonitor and GdkDisplay
functions.

A side effect of this change is that we now use gdk_monitor_get_model
instead of the deprecated gdk_screen_get_monitor_plug_name. It appears
that gdk_monitor_get_model returns similar or same values as the old function
but the documentation is not clear on this. So this might have some side
effects for some users.