tilda_terminal: use asynchronous command spawning
This replaces the synchronous VTE command spawning API with the asynchronous version.

Due to this change error handling and collection of the process id moved to the
'shell_spawned_cb' callback.

If an error with the custom command occured, we now need to spawn the
default shell asynchronously from the 'shell_spawned_cb' callback.
This requires that we have the working directory available in the callback,
so we now store the initial working directory in the tilda_terminal structure,
when we initialize the terminal. This change additionally allows us to also
inherit the working directory when the restart command option is set.

There is now a timeout for shell and command spawning that can be
configured via the 'command_timeout_ms' config option and which defaults
to 3 seconds (3000ms). There is not yet a GUI option for this.

The asynchronous command spawning exposed a bug in the restart command
handling when the child exits. Because both the 'eof' and the 'child-exited'
signals call the 'child_exited_cb' callback the command actually got restarted
twice. With synchronous spawning this was never a problem but with the
asynchronous command spawning it caused a race condition. Therefore the
'eof' signal was removed since the 'child-exited' signal is enough for
restarting commands.

The terminal structure is also now allocated via 'g_new0' to ensure
that all members are properly initialized to zero.