wizard: stop hiding window on monitor switch
We used to hide the window when the monitor
change was updating the spinner widgets.
This was introduced in commit 8f6b8213.

In commit 4c252136 we removed a call to
gtk_window_move which caused a regression
in the monitor switch callback. Originally
we were hiding the window while we moved
and resized it and removal of the gtk_window_move
call broke the ability to move or resize the
hidden window.

Moving the window while its hidden is not
supported by window managers so this change
removes the hide and show calls.