[search] Support deprecated non-pcre2 search
This patch brings back the deprecated GRegex based search
for VTE version before 0.56.1. The reason is that Ubuntu
18.04 and 18.10 have removed pcre2 support from their
respective VTE versions. Using the newer pcre2-based
search functions on these ubuntu versions results in
the runtime error "PCRE2 not available".

This practice was stopped in Ubuntu 19.04. Since
Ubuntu 19.04 ships with VTE 0.56.1 we only enable
the newer pcre2-based search on versions of VTE 0.56.1
or above.

This allows that tilda can still be used on Ubuntu 18.04
and 18.10.

This patch ensures that the latest tilda can still beused
on the current LTS version of Ubuntu. This patch will be
removed some time after Ubuntu 20.04 LTS has been released.

A discussion of this decision can be found at:

* https://github.com/lanoxx/tilda/pull/399