config: remove default for deprecated bold option
The bold option was removed in commit ed4a0a5ff2ef11d06668dcca7b414f22c8a57059
but the default value remained in the config. This prevented
the config value from being removed from the config file and as a result
users would keep seeing the deprecation warning for this value.