tilda_window: use correct titel in switch_page_cb
We should use the tilda_terminal_get_title
function because that computes the whole
title including the configured title
behaviours.

We also need to guard against NULL titles,
this can happen during the initialization
of a new tilda terminal, when the terminal
has not yet been initialized.

Finally, we should also free the title.