tilda_terminal: correctly set title on initial terminal
The check in the page conditional was wrong, the get_current_page function
can also return 0 for the first tab so the condition should check for greater
or equal. Also in case when the note book has no pages
then the current terminal is always considered active and we
should update the window title. This is the case when
there is only one tab and the 'Show single tab'
option is disabled. This happends especially when
tilda has just been started and there is only one
tab.