tilda_terminal: cleanup get_title logic
This moves the title behaviour into the get_title function
so that we can reuse it better across tilda.

We also introduce a new public function called
tilda_terminal_get_full_title which can be used
for the title tooltip of the tab label to display
the full tab title.