key_grabber.c: limit filter function calls
This limits the number of times the filter function is called, when
the user continuously presses the toggle key.

This avoids a BadMatch XError which can happens,
when the user keeps the toggle key pressed for
a longer period of time.

This check should always be made, not only for
certain actions.

Fixes #388