tilda_window: handle switch-page in window
Handling the notebooks "switch-page" signal in the notebook meant that
we were connecting a new signal handler each time a new tab was added.

The signal handlers were never removed, since the notebook will not
be destroyed until the tilda window is closed (e.g. when tilda exits).

The signal handler should instead be connected only once in the
tilda_window so that we have exactly one signal handler registered
independently of how many tabs are open.