tilda-terminal: correctly detach menu
This change correctly detaches the context menu from the tilda window
to ensure that the action groups bound to the menu are correctly
released.

This fixes a bug in tilda's close tab confirmation dialogs, which
caused the confirmation dialog to reappear when the close-tab
hotkey was used and the dialog was closed with the cancel button.

The confirmation dialog was appearing twice, because the close-tab
action is registered both in the context menu's action group as well
as in the action group of the hotkeys. Since the action group
of the context menu was not correctly released the close-tab action
of the context-menu was also being invoked when the close-tab hotkey
was being pressed.