context-menu: correctly free link
This fixes a bug introduced in commit c8f7fc4f which lead to memory not
being freed, when the context menu was closed without using the
"copy-link" action.

This change introduces a separate structure that keeps a reference to
the matched string and ensures that it is freed after the menu
selection has been done.