Ignore BadMatch error with SetInputFocus
This fixes a crash due to the "BadMatch (invalid parameter attributes)"
error when closing the tilda window.

The issue seems linked to a focus event being triggered after the
`gtk_widget_hide` call on the now closed (not viewable) window as
explained in https://tronche.com/gui/x/xlib/input/XSetInputFocus.html

The cause of this issue seems to come from the fact that we register
our own X11 error handler. When we call `gtk_widget_hide` on our main
window, then the window will lose focus and the window manager will
try to focus something else, so it sends a [`WM_TAKE_FOCUS`][1] event
to the application which is handled by GTK and GTK in turn calls
`XSetInputFocus`. The call to XSetInputFocus inside the GTK library
is wrapped in error traps. This means that Gtk normally [ignores][2]
any BadMatch errors that occur as a result of calling `XSetInputFocus`.

However, since we register our own error handler, the `BadMatch` errors
still reaches our own handler. This means that ignoring the `BadMatch`
error in our own handler is actually a valid solution. However,
an even better solution is probably to remove the error handler all
together.

[1]: https://gitlab.gnome.org/GNOME/gtk/-/blob/master/gdk/x11/gdkdisplay-x11.c#L1268
[2]: https://gitlab.gnome.org/GNOME/gtk/-/blob/master/gdk/x11/gdkdisplay-x11.c#L2603

Fixes: #434
Fixes: #431
Fixes: #428
Fixes: #348

Signed-off-by: Sebastian Geiger <sbastig@gmx.net>