tilda_terminal: prepend PATH= to environment value
Prepends "PATH=" to the value returned by getenv("PATH"). This fixes an
assertion error in the VTE library that prevents the terminal from
spawning when using a custom command.

This change is needed due to changed behavior introduced in the VTE
library by [this commit][1]. The old behavior was probably always wrong
and has just never been noticed.

Fixes #435
Fixes #436
Fixes #440

Signed-off-by: Sebastian Geiger <sbastig@gmx.net>

[1]: https://github.com/lanoxx/tilda/commit/47c679fb1fcc8aa6817c089bd97ec2064e04a8f1