tilda_window: add background style class
This change adds a background style class to the main widget container
of the Tilda window and instead removes the hard coded application
style for the search box.

Setting the background style is necessary, because we have configured
the tilda window to be 'app paintable' which removes the implicity
styling of the GTK framework from the window. The consequence is that
in transparent mode not only the terminal widget is transparent but also
all other widget that do not have an explicit background set.

To compensate for this problem we previously added a custom CSS rule
for the search bar background that added a hard-coded white background.
but that did not work well with dark themes.

The change in this issue avoids the hard-coded CSS background and
therefore also works well for dark themes.

Fixes #437