X11: replace error handler by error traps
GTK ignores certain errors when it invokes raw X11 functions, by
registering our own error handler, we may catch some errors that GTK
would have ignored.

For example, when we call `gtk_widget_hide` on our main window in the
key_grabber, then the window will lose focus and the window manager
will try to focus something else, so it sends a [`WM_TAKE_FOCUS`][1]
event to the application which is handled by GTK and GTK in turn calls
`XSetInputFocus`. The call to XSetInputFocus inside the GTK library
is wrapped in error traps. This means that Gtk normally [ignores][2]
any BadMatch errors that occur as a result of calling `XSetInputFocus`.

However, since we register our own error handler, the `BadMatch` errors
still reach our own handler. As a work around we added a check in
commit bbb8582ab4a8d0705de29191bec9af6ac2bb42df to ignore BadMatch errors.

This commit removes the custom error handler completely from tilda. In
addition we add error traps around the relevant X calls that can throw
`BadWindow` or `BadMatch` errors to ensure that these errors are
ignored.

[1]: https://gitlab.gnome.org/GNOME/gtk/-/blob/master/gdk/x11/gdkdisplay-x11.c#L1268
[2]: https://gitlab.gnome.org/GNOME/gtk/-/blob/master/gdk/x11/gdkdisplay-x11.c#L2603
[3]: https://github.com/lanoxx/tilda/commit/3d8fad0219498dc05d408bf0c6df6ac3494bafe6