Allow ergonomic use of `ctrl+c` for copy
This commit makes two small tweaks that allow for sensible use of ctrl+C
to copy:

1. It propagates the "copy" shortcut event if there is no text selected
in the terminal. This allows `ctrl+c` to be forwarded to the underlying
TTY to interrupt the foreground process as per usual.
2. It unselects the current selection when a clipboard copy is made.
This means that after you have used `ctrl+c` to copy the first time it
can be immediately used to interrupt the foreground process without
manually unselecting.