dbus: implement visibility toggler
To toggle visibility in Wayland, commands
have to be sent via dbus.
Tilda has to be started with `--dbus` argument,
which exposes a D-Bus action that toggles the tilda window.
Next, the user is expected to define its own shortcut in their
desktop environment (i.e, Gnome) via `tilda -T 0`.
If more Tilda processes are running, 0 corresponds to D-Bus
acquired name id.

The steps during Tilda startup are changed, given that window creation
is not needed, if we intend to exit right after sending the D-Bus cmd.

Additionally, the argument after the `-T` option is optional if the first
tilda process with instance id 0 should receive the toggle cmd.

Signed-off-by: Krzysztof Królczyk <Krzysztof.Krolczyk@o2.pl>

Signed-off-by: Sebastian Geiger <sbastig@gmx.net>