tilda-lock-files: do not remove the global lock file
If we remove the global lock file, then it may create a race-condition
between tilda processes that are starting concurrently in quick succession.
The reason is that flock locks on the inode and not on the file name. If
the file is deleted and recreated at the start of every process, then it
can happen that the global lock file used in different tilda processes
uses different inodes and then multiple processes can aquire a global
lock at the same time.

This is still not ideal, since it will leave the global lock-file
in the cache directory and there is no way to clean it up. Ultimately,
we should move the locking files into XDG_RUNTIME_DIR to ensure that
it will get cleaned during a reboot.